Separate the code for the menu.
